feat: implement muscle/fat segmentation feature with mask generation

+++ frontend/src/shared/types/viewer.types.ts
@@ -1,5 +1,4 @@
-import type { MeasureResult } from '../api/client'
-import type { FilterType } from '../api/types'
+import type { FilterType, MeasureResult } from '../api'
 
 export type ColormapType = 'gray' | 'jet' | 'hot'
 
@@ -65,7 +64,9 @@ export interface H5Meta {
 }
 
 export interface H5VolumeData extends H5Meta {
-    vIndices: Float32Array
+    // Uint32 (not float32): flat voxel indices reach 32M for a full volume, past
+    // float32's exact-integer limit of 2^24 ≈ 16.7M. See the vertex shader.
+    vIndices: Uint32Array
     vIntensities: Float32Array
     // Uint8Array (values 0-255) instead of Float32Array to cut memory 4×.
     // Divide by 255 when a 0-1 float is needed (e.g. SlicePanel canvas rendering).
@@ -98,7 +99,13 @@ export interface H5TabEntry {
     data: H5VolumeData | null
     hasSlices: boolean
     sourceFile?: File
+    // Dual meaning: for stitched tabs this is the *session* id (drives the
+    // /sessions/{id}/… endpoints); for local files it is the lazily-uploaded
+    // volume id cached by useResolveVolumeId. Consumers distinguish the cases
+    // via sourceFile/registeredVolumeId — see useFilterJob.
     backendVolumeId?: string
+    // Single volume stored server-side under /volumes/{id} (registered by path
+    // or created by crop) — usable by filter/measure/crop/mask endpoints.
     registeredVolumeId?: string
 }
